+++ CHANGELOG.md
@@ -113,6 +113,7 @@
   encoding to `File.read_text`. New `File.read` API.][3390]
 - [Improved the `Range` type. Added a `down_to` counterpart to `up_to` and
   `with_step` allowing to change the range step.][3408]
+- [Aligned `Text.split` API with other methods and added `Text.lines`.][3415]
 
 [debug-shortcuts]:
   https://github.com/enso-org/enso/blob/develop/app/gui/docs/product/shortcuts.md#debug
@@ -173,6 +174,7 @@
 [3393]: https://github.com/enso-org/enso/pull/3393
 [3390]: https://github.com/enso-org/enso/pull/3390
 [3408]: https://github.com/enso-org/enso/pull/3408
+[3415]: https://github.com/enso-org/enso/pull/3415
 
 #### Enso Compiler
 
@@ -183,6 +185,7 @@
 - [Fixed compiler issue related to module cache.][3367]
 - [Fixed execution of defaulted arguments of Atom Constructors][3358]
 - [Converting Enso Date to java.time.LocalDate and back][3374]
+- [Functions with all-defaulted arguments now execute automatically][3414]
 - [Move Builtin Types and Methods definitions to stdlib][3363]
 
 [3227]: https://github.com/enso-org/enso/pull/3227
@@ -193,6 +196,9 @@
 [3360]: https://github.com/enso-org/enso/pull/3360
 [3367]: https://github.com/enso-org/enso/pull/3367
 [3374]: https://github.com/enso-org/enso/pull/3374
+[3412]: https://github.com/enso-org/enso/pull/3412
+[3414]: https://github.com/enso-org/enso/pull/3414
+[3417]: https://github.com/enso-org/enso/pull/3417
 [3363]: https://github.com/enso-org/enso/pull/3363
 
 # Enso 2.0.0-alpha.18 (2021-10-12)